chore: move rdb save related logic into auxillary file
1. No logic was changed during refactoring.
2. Flipped the flag to run regression tests for now own with zset_tree=on
